+++ shiny/src/wrapPMR.R
@@ -1,0 +1,83 @@
+# wraps COL2A1 calibration, PMR calculation and testWIDqEC main work functions, based on user input
+# contact: charlottevavourakis@gmail.com
+# under development for Shiny App GYN test, Last update October 26, 2022
+# Option for using fixed intercept or slope, or external calibration curve, will only be functional in the bare R-pipeline, not in the Shiny apps
+# todo: remove this here?
+
+calculate_pmr <- function(data,
+                          threshold_COL2A1=30,
+                          threshold_targets=38,
+                          external_curve=FALSE,
+                          curve=NULL,
+                          fix_intercept=36.9, #these values come from meta-analysis
+                          fix_slope=-3.4,
+                          calib_fixed=FALSE){
+  
+  ####---- Prepare ----####
+  
+  require(dplyr)
+  require(tidyverse)
+  
+  # main work functions
+  source("../../lib/calibGblock.R")
+  source("../../lib/calcPMRGyn.R")
+  source("../../lib/testWIDqEC.R")
+  
+  # WID-qEC thresholds
+  qEC_threshold1 <- 0.03
+  qEC_threshold2 <- 0.63 #this threshold is not used by Tyrolpath
+  
+  # check that threshold targets >= threshold COL2A1
+  if(threshold_COL2A1>threshold_targets){
+    stop("Error: CT-threshold set for COL2A1 should be lower or equal to that for other targets")
+  }
+  
+  # collect target and sample overviews
+  targets <- unique(data$Target)
+  samples <- unique(data$Sample)
+  
+  # initialize list so multiple objects can be returned from function 
+  list_out <- list()
+  
+  ####---- Calibrate COL2A1 ----####
+  
+  calibration <- calibGblock(data)
+  
+  # add plot to list
+  list_out[[1]] <- calibration[[1]]
+  
+  # If user choose fixed values overwrite the coefficients from standard curve
+  if(calib_fixed==TRUE){
+    intercept <- fix_intercept
+    slope <- fix_slope
+  } else{
+    intercept <- calibration[[2]]
+    slope <- calibration[[3]]
+  }
+  
+  ####---- Reformat data ~reps knowing plate layout  ----####
+  
+  data <- data %>%
+    mutate(rep = case_when(grepl("A|C|E|G|I|K|M|O", Well.Position) == TRUE ~ 1,
+                           grepl("B|D|F|H|J|L|N|P", Well.Position) == TRUE ~ 2))
+  
+  ####---- Calculate PMR ----####
+  # based on default or user set thresholds
+  # minimum two reps needed, not sure whether works for more
+  
+  calcPMR <- calcPMRGyn(data,targets,samples,threshold_COL2A1, threshold_targets)
+  list_out <- append(list_out,calcPMR)
+  
+  ####---- Make the final summary commercial WIDqEC test ----####
+  
+  targets_qEC <- c("GYPC1","GYPC2","ZSCAN12")
+  if(all(targets_qEC %in% targets)){
+    final <- testWIDqEC(calcPMR[[1]],samples,calcPMR[[5]],calcPMR[[6]],calcPMR[[7]])
+    list_out[[9]] <- final #information in final csv file
+  }
+  
+  ####---- Return or save results ----####
+  
+  return(list_out)
+  
+}